requirements.txt
Django==4.2.3
djangorestframework==3.14.0
ping3==4.0.4
django-extensions==3.2.3
